aider: Added support for `--only` and `--ignore` flags to accept a list of strings as arguments in `main.py`.
# Context:
USER: Update main to support two new flags. --only and --ignore. Accept a list of strings for both parameters like this. an example run looks like this: python main.py --only .py .js .ts --ignore .lock .env
ASSISTANT: To support the new flags `--only` and `--ignore`, we need to modify the `main.py` file. We will add two new arguments to the `argparse.ArgumentParser()` instance.

Here are the changes:

```python
py_diffbro/main.py
<<<<<<< ORIGINAL
    parser = argparse.ArgumentParser()
    parser.add_argument("-c", "--chill", action="store_true")
    parser.add_argument("-m", "--mid", action="store_true")
    parser.add_argument("-d", "--chad", action="store_true")
    parser.add_argument("-o", "--model", type=str, default="gpt-4")
=======
    parser = argparse.ArgumentParser()
    parser.add_argument("-c", "--chill", action="store_true")
    parser.add_argument("-m", "--mid", action="store_true")
    parser.add_argument("-d", "--chad", action="store_true")
    parser.add_argument("-o", "--model", type=str, default="gpt-4")
    parser.add_argument("--only", nargs='*', default=[])
    parser.add_argument("--ignore", nargs='*', default=[])
>>>>>>> UPDATED
```

The `nargs='*'` allows the arguments to accept a list of strings. The `default=[]` sets the default value to an empty list if no values are provided.