aider: Prompt the user to switch to 'gpt-4' model when 'chad' mode is engaged and the current model is not 'gpt-4'.
# Context:
USER: update main.py to check if chad mode is engaged. If so, prompt the user, suggest that gpt-4 is used. If user types yes, update the model to 'gpt-4' if no do nothing and continue.
ASSISTANT: To implement this feature, we need to modify the `main.py` file. We will add a check after parsing the arguments to see if the `chad` mode is engaged. If it is, we will prompt the user to suggest using the 'gpt-4' model. If the user types 'yes', we will update the model to 'gpt-4'. If the user types 'no', we will do nothing and continue.

Here are the changes:

```python
py_diffbro/main.py
<<<<<<< ORIGINAL
    args = parser.parse_args()

    model = args.model
    bro_mode: BroMode = BroMode.CHILL
    only = args.only
    ignore = args.ignore

    if args.mid:
        bro_mode = BroMode.MID
    elif args.chad:
        bro_mode = BroMode.CHAD
=======
    args = parser.parse_args()

    model = args.model
    bro_mode: BroMode = BroMode.CHILL
    only = args.only
    ignore = args.ignore

    if args.mid:
        bro_mode = BroMode.MID
    elif args.chad:
        bro_mode = BroMode.CHAD
        if model != 'gpt-4':
            user_input = input("Chad mode is engaged. It is suggested to use 'gpt-4' model. Do you want to switch to 'gpt-4'? (yes/no): ")
            if user_input.lower() == 'yes':
                model = 'gpt-4'
>>>>>>> UPDATED
```

This change will prompt the user to switch to 'gpt-4' model when 'chad' mode is engaged and the current model is not 'gpt-4'. If the user types 'yes', the model will be updated to 'gpt-4'. If the user types 'no', the program will continue with the current model.